+++ lessons/l30-entities/src/main.c
@@ -69,7 +69,7 @@ int main(void)
     asset_init_compression(2);
     dfs_init(DFS_DEFAULT_LOCATION);
     display_init(RESOLUTION_320x240, DEPTH_16_BPP, FB_COUNT, GAMMA_NONE,
-                 FILTERS_RESAMPLE_ANTIALIAS);
+                 FILTERS_RESAMPLE);
     rdpq_init();
     joypad_init();
     rdpq_text_register_font(1, rdpq_font_load_builtin(FONT_BUILTIN_DEBUG_VAR));
@@ -226,7 +226,7 @@ int main(void)
         rdpq_attach(display_get(), display_get_zbuf());
         t3d_frame_start();
         t3d_viewport_attach(&viewport);
-        t3d_screen_clear_color(RGBA32(30, 55, 95, 0));
+        t3d_screen_clear_color(RGBA32(30, 55, 95, 0xFF));
         t3d_screen_clear_depth();
         t3d_light_set_ambient(amb);
         t3d_light_set_directional(0, dirC, &ldir);