+++ lessons/l34-juice/src/main.c
@@ -68,7 +68,7 @@ int main(void)
     asset_init_compression(2);
     dfs_init(DFS_DEFAULT_LOCATION);
     display_init(RESOLUTION_320x240, DEPTH_16_BPP, FB_COUNT, GAMMA_NONE,
-                 FILTERS_RESAMPLE_ANTIALIAS);
+                 FILTERS_RESAMPLE);
     rdpq_init();
     joypad_init();
     rdpq_text_register_font(1, rdpq_font_load_builtin(FONT_BUILTIN_DEBUG_VAR));
@@ -303,10 +303,10 @@ int main(void)
         t3d_frame_start();
         t3d_viewport_attach(&viewport);
 
-        color_t clear = RGBA32(25, 55, 100, 0);
+        color_t clear = RGBA32(25, 55, 100, 0xFF);
         if (flash > 0.f) {
             uint8_t add = (uint8_t)(flash * 80.f);
-            clear = RGBA32(25 + add, 55 + add / 2, 100, 0);
+            clear = RGBA32(25 + add, 55 + add / 2, 100, 0xFF);
         }
         t3d_screen_clear_color(clear);
         t3d_screen_clear_depth();