+++ lessons/m3-asset-lab/src/main.c
@@ -54,7 +54,7 @@ int main(void)
 
     dfs_init(DFS_DEFAULT_LOCATION);
     display_init(RESOLUTION_320x240, DEPTH_16_BPP, FB_COUNT, GAMMA_NONE,
-                 FILTERS_RESAMPLE_ANTIALIAS);
+                 FILTERS_RESAMPLE);
     rdpq_init();
     joypad_init();
     rdpq_text_register_font(1, rdpq_font_load_builtin(FONT_BUILTIN_DEBUG_VAR));
@@ -174,7 +174,7 @@ int main(void)
         rdpq_attach(display_get(), display_get_zbuf());
         t3d_frame_start();
         t3d_viewport_attach(&viewport);
-        t3d_screen_clear_color(RGBA32(35, 45, 70, 0));
+        t3d_screen_clear_color(RGBA32(35, 45, 70, 0xFF));
         t3d_screen_clear_depth();
 
         t3d_light_set_ambient(ambient);